channel_service/app/core/telegram_bot.py
import os
from telegram import Bot
from dotenv import load_dotenv

load_dotenv()

TELEGRAM_BOT_TOKEN = os.getenv("TELEGRAM_BOT_TOKEN")

def get_bot() -> Bot:
    if not TELEGRAM_BOT_TOKEN:
        raise ValueError("TELEGRAM_BOT_TOKEN no está configurado en el entorno.")
    return Bot(token=TELEGRAM_BOT_TOKEN)