channel_service/requirements.txt
fastapi
uvicorn
python-telegram-bot
requests
python-dotenv